FIX: Remove copy, move testing to _testing
